getApps and getStudies Super Admin changes and study metadata api changes
getApps and getStudies Super Admin changes and study metadata changes

+++ participant-manager-module/participant-manager-service/src/main/java/com/google/cloud/healthcare/fdamystudies/beans/AppResponse.java
@@ -23,9 +23,9 @@ public class AppResponse extends BaseResponse {
 
   private List<AppDetails> apps = new ArrayList<>();
 
-  private int studyPermissionCount;
+  private long studyPermissionCount;
 
-  public AppResponse(MessageCode messageCode, List<AppDetails> apps, int studyPermissionCount) {
+  public AppResponse(MessageCode messageCode, List<AppDetails> apps, long studyPermissionCount) {
     super(messageCode);
     this.apps.addAll(apps);
     this.studyPermissionCount = studyPermissionCount;
